Move step logic back into a single file.
